Add authentication method in request scope (#2573)
Introduce field AuthMethod in request scope.

Allow middlewares to store the authentication method in AuthMethod
field of the RequestScope. It is necessary for the proxy to be aware
if authentication was achieved through a cookie or a header,
as only requests that are authenticated via cookie need CSRF protection,
while programmatic client requests that use an authorization header
do not need validation of CSRF tokens.

Co-authored-by: Athina Plaskasoviti <athina.plaskasoviti@hpe.com>
Co-authored-by: Konstantinos Lolos <konstantinos.lolos@hpe.com>