Implement AuthMethod header for upstreams (#2573)
Allow forwarding the authentication method from the request scope
to upstream. This is necessary in case the CSRF token validation
is implemented on the upstream, and not necessary to enable on the
proxy, so upstream can decide when to apply or skip validation,
since only requests that are authenticated by cookie should be
subjected to CSRF protection.

Co-authored-by: Athina Plaskasoviti <athina.plaskasoviti@hpe.com>
Co-authored-by: Konstantinos Lolos <konstantinos.lolos@hpe.com>